+++ validation/handler.ts
@@ -4,23 +4,24 @@ import { z } from "zod";
 
 /**
  * A wrapper for handling validation of POST requests.
- * Ensures both the body and response objects follow the expected schema.
+ * Ensures both the request and response objects will follow the expected schema,
+ * returns error status codes if not.
  *
  * @type T is the request body type,
  * @type U is the expected response type,
  * @type V is the optional transformed response type.
  *
- * @param req body of the request
+ * @param req the request
  * @param bodyValidator zod object used to validate the body
  * @param responseValidator zod object used to validate the response
  * @param fetchUrlExtractor method used to get the fetch url
- * @param responseTransformer optional method used to alter the default response object
+ * @param responseTransformer optional method used to alter the default response object, e.g. flattening or only returning a specified part
  */
 export async function postValidationHandler<T, U, V = any>(
   req: NextRequest,
   bodyValidator: z.AnyZodObject,
   responseValidator: z.AnyZodObject | z.ZodArray<z.AnyZodObject>,
-  fetchUrlExtractor: (body: T) => string,
+  fetchUrlExtractor: (body: T, ...args: any) => string,
   responseTransformer?: (body: T, response: U, ...args: any) => V
 ) {
   const body: T = await req.json();
@@ -67,18 +68,19 @@ export async function postValidationHandler<T, U, V = any>(
 
 /**
  * A wrapper for handling GET requests.
- * Ensures the response follows the expected schema.
+ * Ensures the response object will follow the expected schema,
+ * returns error status codes if not.
  *
  * @type T is the expected response type,
  * @type U is the optional transformed response type.
  *
  * @param responseValidator zod object used to validate structure of response
  * @param fetchUrlExtractor method used to get the url
- * @param responseTransformer optional method for transforming the format of the response, takes response as input
+ * @param responseTransformer optional method for transforming the format of the response
  */
 export async function getValidationHandler<T, U = any>(
   responseValidator: z.AnyZodObject | z.ZodArray<z.AnyZodObject>,
-  fetchUrlExtractor: () => string,
+  fetchUrlExtractor: (...args: any) => string,
   responseTransformer?: (response: T, ...args: any) => U
 ) {
   const fetchURL = fetchUrlExtractor();
